Refactor aws-janitor into a bunch of packages instead of one giant one
