+++ maintenance/aws-janitor/BUILD.bazel
@@ -1,23 +1,17 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "elb.go",
-        "main.go",
-    ],
+    srcs = ["main.go"],
     importpath = "k8s.io/test-infra/maintenance/aws-janitor",
     visibility = ["//visibility:private"],
     deps = [
+        "//maintenance/aws-janitor/resources:go_default_library",
+        "//maintenance/aws-janitor/s3:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
-        "//vendor/github.com/aws/aws-sdk-go/aws/awserr:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/aws/session:go_default_library",
-        "//vendor/github.com/aws/aws-sdk-go/service/autoscaling:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/service/ec2:go_default_library",
-        "//vendor/github.com/aws/aws-sdk-go/service/elb:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/service/iam:go_default_library",
-        "//vendor/github.com/aws/aws-sdk-go/service/route53:go_default_library",
-        "//vendor/github.com/aws/aws-sdk-go/service/s3:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
     ],
 )
@@ -37,18 +31,11 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [":package-srcs"],
+    srcs = [
+        ":package-srcs",
+        "//maintenance/aws-janitor/resources:all-srcs",
+        "//maintenance/aws-janitor/s3:all-srcs",
+    ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    size = "small",
-    srcs = ["janitor_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
-        "//vendor/github.com/aws/aws-sdk-go/service/route53:go_default_library",
-    ],
 )
