+++ maintenance/aws-janitor/resources/elb.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2018 The Kubernetes Authors.
+Copyright 2019 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package resources
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
@@ -25,9 +25,9 @@ import (
 
 // Clean-up ELBs
 
-type loadBalancers struct{}
+type LoadBalancers struct{}
 
-func (loadBalancers) MarkAndSweep(sess *session.Session, account string, region string, set *awsResourceSet) error {
+func (LoadBalancers) MarkAndSweep(sess *session.Session, account string, region string, set *Set) error {
 	svc := elb.New(sess, &aws.Config{Region: aws.String(region)})
 
 	var toDelete []*loadBalancer // Paged call, defer deletion until we have the whole list.
