+++ maintenance/aws-janitor/resources/subnets
@@ -1,0 +1,1 @@
+package resources
